1060519005569433690/requirements.txt
async_class==0.5.0
asyncpraw==7.5.0
discord.py==2.1.0